Variables
Implemented variables

+++ subjects/Time.py
@@ -1,5 +1,5 @@
 from time import sleep
 
-wait = lambda a, p: sleep(int(a))
+wait = lambda a, p: sleep(float(a))
 
 functions = {"wait":wait}
